ItemMover Reservation
Item movers now store a first-come-first-serve queue of reservations instead of a single one.

This allows for multiple items to wait in line for a slot. i.e. when a bottleneck occurs, the inserter and belt will take turns inserting into next slot, instead of only the first one getting the opportunity every single time.